update all zekr screen add  a search bar for it
