+++ app/src/main/java/com/example/anees/ui/screens/azkar/azkar_details.kt
@@ -26,6 +26,7 @@ import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
 import com.example.anees.data.model.ZekrModelItem
 import com.example.anees.ui.screens.azkar.component.DetailsCard
+import com.example.anees.utils.AzkarUtils
 import kotlinx.coroutines.delay
 
 @Composable
