+++ app/src/main/java/com/example/anees/ui/screens/sebha/component/azkar_grid_view.kt
@@ -9,11 +9,11 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
 import com.example.anees.data.model.Zekir
+import  com.example.anees.utils.azkarList
 
 
 @Composable
 fun ZekrHorizontalStaggeredGrid(
-    azkarList: List<Zekir>,
     currentIndex: Int,
     onZekrClick: (Int) -> Unit
 ) {
